Feat: Add language info and description to search list row
